ai-service/requirements.txt
accelerate==0.21.0
aenum==3.1.15
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
async-timeout==5.0.1
attrs==25.1.0
blendmodes==2022
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
clean-fid==0.1.35
click==8.1.8
clip @ https://github.com/openai/CLIP/archive/d50d76daa670286dd6cacf3bcd80b5e4823fc8e1.zip#sha256=b5842c25da441d6c581b53a5c60e0c2127ebafe0f746f8e15561a006c6c3be6a
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
deprecation==2.1.0
diffusers==0.35.1
diskcache==5.6.3
einops==0.4.1
exceptiongroup==1.2.2
facexlib==0.3.0
fastapi==0.94.0
ffmpy==0.5.0
filelock==3.17.0
filterpy==1.4.5
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.32
gradio==3.41.2
gradio_client==0.5.0
h11==0.12.0
httpcore==0.15.0
httpx==0.24.1
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.5
jsonmerge==1.8.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
kornia==0.6.7
lark==1.1.2
lazy_loader==0.4
lightning-utilities==0.12.0
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.1.0
narwhals==1.28.0
networkx==3.4.2
numba==0.61.0
numpy==1.26.2
omegaconf==2.2.3
open-clip-torch==2.20.0
openai-whisper==20240930
opencv-python==4.11.0.86
orjson==3.10.15
packaging==24.2
pandas==2.2.3
piexif==1.1.3
Pillow==9.5.0
pillow-avif-plugin==1.4.3
propcache==0.3.0
protobuf==3.20.0
psutil==5.9.5
pydantic==1.10.21
pydantic_core==2.27.2
pydub==0.25.1
pygame==2.6.1
Pygments==2.19.1
pynput==1.7.7
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==1.9.4
pytz==2025.1
PyWavelets==1.8.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
resize-right==0.0.2
rich==13.9.4
rpds-py==0.23.1
ruff==0.9.7
safehttpx==0.1.6
safetensors==0.6.2
scikit-image==0.21.0
scipy==1.15.2
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
spandrel==0.3.4
spandrel_extra_arches==0.1.1
starlette==0.26.1
sympy==1.13.1
tifffile==2025.2.18
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.22.0
tomesd==0.1.3
tomlkit==0.13.2
torch==2.5.1+cu121
torchdiffeq==0.2.3
torchmetrics==1.6.1
torchsde==0.2.6
torchvision==0.20.1+cu121
tqdm==4.67.1
trampoline==0.1.2
transformers==4.56.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uv==0.8.15
uvicorn==0.34.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.1.3
yarl==1.18.3
zipp==3.23.0
